feature: project 14 - first commit

+++ controllers/card.js
@@ -1,46 +1,55 @@
 const Card = require('../models/card');
 const ErrorNotFound = require('../errors/ErrorNotFound');
+const ErrorForbidden = require('../errors/ErrorForbidden');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send({ data: cards }))
-    .catch((err) => res.status(500).send({ message: err.message }));
+    .catch(next);
 };
 
-module.exports.createCard = (req, res) => {
-  const { name, link } = req.body;
-  Card.create({ name, link, owner: req.user._id })
+module.exports.createCard = (req, res, next) => {
+  const {
+    name,
+    link,
+  } = req.body;
+  Card.create({
+    name,
+    link,
+    owner: req.user._id,
+  })
     .then((card) => res.send({
-      name: card.name, link: card.link, owner: card.owner, likes: card.likes, _id: card._id,
+      name: card.name,
+      link: card.link,
+      owner: card.owner,
+      likes: card.likes,
+      _id: card._id,
     }))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        return res.status(400).send({ message: 'Переданы некорректные данные при создании карточки' });
-      }
-      return res.status(500).send({ message: err.message });
-    });
+    .catch(next);
 };
 
-module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndDelete(req.params.cardId)
+module.exports.deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .orFail(() => {
       throw new ErrorNotFound(`Карточки с id:${req.params.cardId} не существует`);
     })
-    .then((card) => res.send({
-      name: card.name, link: card.link, owner: card.owner, likes: card.likes, _id: card._id,
-    }))
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(400).send({ message: 'Передан неккоректный id карточки' });
+    .then((user) => {
+      if (user.owner.toString() !== req.user._id) {
+        throw new ErrorForbidden('Вы не можете удалять чужие карточки');
       }
-      if (err.statusCode === 404) {
-        return res.status(404).send({ message: `Карточка с _id:${req.params.cardId} не найдена` });
-      }
-      return res.status(500).send({ message: err.message });
-    });
+      Card.findByIdAndDelete(req.params.cardId)
+        .then((card) => res.send({
+          name: card.name,
+          link: card.link,
+          owner: card.owner,
+          likes: card.likes,
+          _id: card._id,
+        }));
+    })
+    .catch(next);
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -50,20 +59,16 @@ module.exports.likeCard = (req, res) => {
       throw new ErrorNotFound(`Карточки с id:${req.params.cardId} не существует`);
     })
     .then((card) => res.send({
-      name: card.name, link: card.link, owner: card.owner, likes: card.likes, _id: card._id,
+      name: card.name,
+      link: card.link,
+      owner: card.owner,
+      likes: card.likes,
+      _id: card._id,
     }))
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(400).send({ message: 'Переданы некорректные данные для постановки лайка.' });
-      }
-      if (err.statusCode === 404) {
-        return res.status(404).send({ message: `Карточка с _id:${req.params.cardId} не найдена` });
-      }
-      return res.status(500).send({ message: err.message });
-    });
+    .catch(next);
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -73,15 +78,11 @@ module.exports.dislikeCard = (req, res) => {
       throw new ErrorNotFound(`Карточки с id:${req.params.cardId} не существует`);
     })
     .then((card) => res.send({
-      name: card.name, link: card.link, owner: card.owner, likes: card.likes, _id: card._id,
+      name: card.name,
+      link: card.link,
+      owner: card.owner,
+      likes: card.likes,
+      _id: card._id,
     }))
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(400).send({ message: 'Переданы некорректные данные для снятия лайка.' });
-      }
-      if (err.statusCode === 404) {
-        return res.status(404).send({ message: `Карточка с _id:${req.params.cardId} не найдена` });
-      }
-      return res.status(500).send({ message: err.message });
-    });
+    .catch(next);
 };
